v.0.16.2.1 Додали decimalInput, поки без перевірки TODO: підключити перевірку (v. 0.16.2.2)
Перед commit виконуємо три команди:
dart format .
flutter analyze
dart run build_runner build --delete-conflicting-outputs

+++ lib/features/learning/data/fixed_task_data/fixed_data_addition.dart
@@ -30,7 +30,7 @@ abstract final class AdditionTaskData {
     InputTaskData(
       prompt: TaskPrompt(content: [TextContent('Скільки буде 2 + 6?')]),
       correctAnswer: 8,
-      inputMode: InputMode.integer,
+      inputMode: InputMode.decimal,
     ),
 
     InputTaskData(
